Make standard_case hash global to allow adding XS easily:
In order to implements parts of HTTP::Headers::Fast in XS, it needs
access to the %standard_case hash which is used by other variables.

It's possible to reimplement that but it's easier to simply access
the available Perl-side existing hash.

Making the %standard_case a global would make it easier for the XS
to reach it without playing with the scratchpad.

For what it's worth, there are already some functions implemented
in XS assuming that this variable is available.